+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Support for the `X-Error-Id` header for tracing purposes
+* Support for relative param in redirect, forcing relative redirects
 
 ### Changed